+++ package.json
@@ -19,26 +19,26 @@
     "react-dropzone": "3.12.2"
   },
   "devDependencies": {
-    "@kadira/storybook": "2.35.3",
-    "@kadira/storybook-deployer": "1.2.0",
+    "@kadira/storybook": "^2.35.3",
+    "@kadira/storybook-deployer": "^1.2.0",
     "autoprefixer-stylus": "^0.13.0",
-    "babel-cli": "6.24.0",
-    "babel-eslint": "7.1.1",
-    "babel-plugin-transform-class-properties": "6.23.0",
-    "babel-plugin-transform-object-rest-spread": "6.23.0",
-    "babel-plugin-transform-react-jsx": "6.23.0",
-    "babel-preset-es2015": "6.24.0",
-    "eslint": "3.16.1",
-    "eslint-config-react-app": "0.6.2",
-    "eslint-loader": "1.6.3",
-    "eslint-plugin-flowtype": "2.21.0",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-plugin-jsx-a11y": "4.0.0",
-    "eslint-plugin-react": "6.4.1",
+    "babel-cli": "^6.24.0",
+    "babel-eslint": "^7.1.1",
+    "babel-plugin-transform-class-properties": "^6.23.0",
+    "babel-plugin-transform-object-rest-spread": "^6.23.0",
+    "babel-plugin-transform-react-jsx": "^6.23.0",
+    "babel-preset-es2015": "^6.24.0",
+    "eslint": "^3.16.1",
+    "eslint-config-react-app": "^0.6.2",
+    "eslint-loader": "^1.6.3",
+    "eslint-plugin-flowtype": "^2.21.0",
+    "eslint-plugin-import": "^2.0.1",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "^6.4.1",
     "husky": "^0.13.2",
-    "lint-staged": "3.4.0",
-    "prettier": "0.22.0",
-    "stylus": "0.54.5"
+    "lint-staged": "^3.4.0",
+    "prettier": "^0.22.0",
+    "stylus": "^0.54.5"
   },
   "peerDependencies": {
     "react": "^15.0.0",
